ai-meeting-notes/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
transformers
torch
ffmpeg-python
sqlalchemy

